new_panel/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[argon2]
python-multipart
python-socketio
aiofiles
requests
pydantic-settings
